Moved Cover Image edit function into separate file
