+++ test/integration/full-content/fixtures/core__cover-image.serialized.html
@@ -1,3 +1,3 @@
-<!-- wp:cover-image {"url":"https://cldup.com/uuUqE_dXzy.jpg","dimRatio":40} -->
+<!-- wp:cover-image {"dimRatio":40,"url":"https://cldup.com/uuUqE_dXzy.jpg"} -->
 <div class="wp-block-cover-image has-background-dim-40 has-background-dim" style="background-image:url(https://cldup.com/uuUqE_dXzy.jpg)"><p class="wp-block-cover-image-text">Guten Berg!</p></div>
 <!-- /wp:cover-image -->
